Atualizado CRUDUser para persistir user e nao userGroup Add alguns try/catch Comentado alguns metodos nao utilizados Eliminado alguns blocos de comentarios

+++ easybpms/src/main/java/com/easybpms/domain/IDomainEntity.java
@@ -2,7 +2,7 @@ package com.easybpms.domain;
 
 public interface IDomainEntity {
 	
-	public int getId();
+	public long getId();
 	public String getTenancy();
 
 }
